Search: index from html files for mkdocs projects
It's under a feature flag currently,
since I would like to test this with sphinx later.

Tested with all themes available in
https://github.com/mkdocs/mkdocs/wiki/MkDocs-Themes

I did a search in all of them to check that sections are correctly indexed,
and that irrelevant content isn't indexed.

I wasn't able to build these,
but I checked the html, and it's very similar to the default mkdocs theme,
so it should work.

- https://github.com/michaeltlombardi/mkdocs-psinder
- https://github.com/byrnereese/mkdocs-bootstrap4
- https://github.com/daizutabi/mkdocs-ivory

These themes don't expose search,
but the content is indexed and we show results in the dashboard.

- https://gitlab.com/lramage/mkdocs-bootstrap386
- https://gitlab.com/lramage/mkdocs-gitbook-theme

Things to do later:

Write a guide with recommendations for static site generators/themes
authors, to integrate nice with Read the Docs.

Other problems I found:

We don't override the search for https://github.com/squidfunk/mkdocs-material
(they don't use the identifier we use to override search).
That isn't related to this, but just wanted to mention it.
We can send a patch to fix that.

You can check the results by either building mkdocs projects and
searching (you need to enable the mkdocs search feature flag and the
index from html feature flag), or just by opening the html files and
checking the content of the json files from tests and see if the indexed
content makes sense (note that I remove some content from the original
html file to make it easier to read the json files, since json doesn't
support multiple lines...)